# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/uoms:
    get:
      tags:
        - Uoms
      summary: List Uoms
      description: Get a paginated list of Uoms with optional filtering and sorting
      operationId: listUoms
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UomList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Uoms
      summary: Create Uom
      description: Create a new Uom
      operationId: createUom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUomRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Uom'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Uom:
      type: object
      required:
        - id
        - company_id
        - code
        - name
        - uom_type
        - decimal_places
        - factor
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 20
        name:
          type: string
          maxLength: 80
        uom_type:
          type: string
        decimal_places:
          type: integer
          format: int32
          minimum: 0
          maximum: 6
        relative_uom_id:
          nullable: true
          type: string
          format: uuid
        relative_factor:
          nullable: true
          type: string
        factor:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateUomRequest:
      type: object
      required:
        - company_id
        - code
        - name
        - uom_type
        - decimal_places
        - factor
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 20
        name:
          type: string
          maxLength: 80
        uom_type:
          type: string
        decimal_places:
          type: integer
          format: int32
          minimum: 0
          maximum: 6
        relative_uom_id:
          nullable: true
          type: string
          format: uuid
        relative_factor:
          nullable: true
          type: string
        factor:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    UomList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Uom'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []